+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 History
 -------
 
+ - New: option single_instance in job configuration
+ - New: logging can be disabled in application or job configuration
+ - Add function set_sigterm_handler() to API
+ - Upgrade dependency: salmagundi -> 0.11.2
+ - Bugfix: error when sending email if file list contained source or target errors
+
 **2020-01-09 (0.8.0)**
  - Emails are now customizable
  - Upgrade dependencies: paramiko -> 2.7.1, salmagundi -> 0.10.0